requirements/base.txt
Django>=1.5.2

dj-static>=0.0.5
gunicorn>=17.5
psycopg2>=2.5
static>=0.4
boto>=2.9.9

dj-database-url>=0.2.1

bpython>=0.12
django-braces>=1.0.0
django-model-utils>=1.3.1
logutils>=0.3.3
South>=0.7.6

django-cors-headers>=0.11

django-extensions>=1.1.1
django-extra-views>=0.6.2
django-floppyforms>=1.1

django-angular>=0.1.4

djangorestframework>=2.3.6
django-filter>=0.6
django-rest-swagger>=0.1.3
django-oauth2-provider>=0.2.6
